Remove Activator class, rename spree_activators to spree_promotions
