+++ core/db/migrate/20130813232134_rename_activators_to_promotions.rb
@@ -1,0 +1,5 @@
+class RenameActivatorsToPromotions < ActiveRecord::Migration
+  def change
+    rename_table :spree_activators, :spree_promotions
+  end
+end
